# Store alphas as `u8` instead of explicitly packing