--- a/sparse_strips/vello_common/src/coarse.rs
+++ b/sparse_strips/vello_common/src/coarse.rs
@@ -114,9 +114,10 @@ impl Wide {
             let x0 = strip.x;
             let strip_y = strip.strip_y();
             let mut col = strip.alpha_idx / u32::from(Tile::HEIGHT);
-            // Can potentially be 0, if the next strip's x values is also < 0.
-            let strip_width =
-                (next_strip.alpha_idx / u32::from(Tile::HEIGHT)).saturating_sub(col) as u16;
+            let next_col = next_strip.alpha_idx / u32::from(Tile::HEIGHT);
+            // Can potentially be 0, e.g. if the strip only changes coarse winding, but this
+            // depends on exact details of strip footprints.
+            let strip_width = next_col.saturating_sub(col) as u16;
             let x1 = x0 + strip_width;
             let tile_x0 = x0 / WideTile::WIDTH;
             // It's possible that a strip extends into a new wide tile, but we don't actually