This stops packing alpha values explicitly into `u32`s, which was a leftover from using `u32` GPU alpha storage buffers. This simplifies the code a bit and is a performance improvement. On my machine, 500 iterations of Paris-30k measure as the following.

```
Strip rendering
    before: 29.770845ms
    after:  28.914503ms
CPU fine
    before: 13.367204ms
    after:  12.358518ms
```